Rework demo using partial
