+++ packages/dx-react-grid-demos/src/demo-sources/grid-editing/inline-cell-editing.partial.jsxt
@@ -1,5 +1,4 @@
-import React, { useState } from 'react';
-import Paper from '@material-ui/core/Paper';
+import React, { useState } from 'react';<%&additionalImports%>
 import { EditingState } from '@devexpress/dx-react-grid';
 import {
   Grid,
@@ -7,12 +6,13 @@ import {
   TableHeaderRow,
   TableInlineCellEditing,
   Toolbar,
-} from '@devexpress/dx-react-grid-bootstrap3';
+} from '@devexpress/dx-react-grid-material-ui';
 import {
   Plugin,
   Template,
   TemplatePlaceholder,
 } from '@devexpress/dx-react-core';
+<%&imports%>
 
 import {
   generateRows,
@@ -21,57 +21,7 @@ import {
 
 const getRowId = row => row.id;
 
-// #FOLD_BLOCK
-const StartEditActionSelector = (props) => {
-  const { defaultAction, changeAction } = props;
-  return (
-    <div>
-      {'Start Edit Action:'}
-      &nbsp;
-      <select
-        defaultValue={defaultAction}
-        onChange={e => changeAction(e.target.value)}
-        className="dropdown"
-      >
-        <option value="click">Click</option>
-        <option value="doubleClick">Double Click</option>
-      </select>
-    </div>
-  );
-};
-
-// #FOLD_BLOCK
-const SelectTextChecker = (props) => {
-  const { isSelectText, changeSelectText } = props;
-  return (
-    <div
-      className="checkbox"
-    >
-      <label
-        htmlFor="selectTextChecker"
-      >
-        <input
-          type="checkbox"
-          checked={isSelectText}
-          id="selectTextChecker"
-          name="selectTextChecker"
-          onChange={e => changeSelectText(e.target.checked)}
-        />
-        Select Text On Focus
-      </label>
-    </div>
-  );
-};
-
-const EditPropsPanel = props => (
-  <Plugin name="EditPropsPanel">
-    <Template name="toolbarContent">
-      <SelectTextChecker {...props} />
-      <TemplatePlaceholder />
-      <StartEditActionSelector {...props} />
-    </Template>
-  </Plugin>
-);
+<%&body%>
 
 export default () => {
   const [columns] = useState([
@@ -110,7 +60,7 @@ export default () => {
   };
 
   return (
-    <Paper>
+    <<%&wrapperTag%><%&wrapperAttributes%>>
       <Grid
         rows={rows}
         columns={columns}
@@ -131,6 +81,6 @@ export default () => {
           selectTextOnEditStart={selectTextOnEditStart}
         />
       </Grid>
-    </Paper>
+    </<%&wrapperTag%>>
   );
 };
